Rename `BaseDispatcher` to `InMemoryDispatcher` to clarify how it works
