+++ src/dispatcher/__init__.py
@@ -1,8 +1,8 @@
 from .ABC import AsyncDispatcher, context, Dispatcher, STOP_SIGNAL
-from .async_base_dispatcher import AsyncBaseDispatcher
+from .async_in_memory_dispatcher import AsyncInMemoryDispatcher
 from .async_amqp_dispatcher import AsyncAMQPDispatcher
 from .async_redis_dispatcher import AsyncRedisDispatcher
-from .base_dispatcher import BaseDispatcher
+from .in_memory_dispatcher import InMemoryDispatcher
 from .context_var_wrapper import ContextVarWrapper
 from .event_handler import AsyncEventHandler, EventHandler
 from .exceptions import StopEvent, UnknownEvent
