+++ src/dispatcher/in_memory_dispatcher.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 import logging
 
-from ._pubsub import AsyncPubSub
-from .ABC import AsyncDispatcher
+from ._pubsub import StupidPubSub
+from .ABC import Dispatcher
 
 
-class AsyncBaseDispatcher(AsyncDispatcher):
+class InMemoryDispatcher(Dispatcher):
     """A simple in memory Pub Sub-based event dispatcher
 
     This class implements an event dispatcher using StupidPubSub as the message
@@ -15,6 +15,8 @@ class AsyncBaseDispatcher(AsyncDispatcher):
 
     :param namespace: The name of the dispatcher the events will be sent from
                       and sent to
+    :param pubsub: A pub sub having at least the methods 'listen', 'publish' and
+                   'subscribe'
     :param parent_logger: A logging.Logger instance. The dispatcher logger
                           will be set to 'parent_logger.namespace'
     """
@@ -23,30 +25,21 @@ class AsyncBaseDispatcher(AsyncDispatcher):
             namespace: str,
             parent_logger: logging.Logger = None
     ) -> None:
-        self.pubsub = AsyncPubSub()
+        self.pubsub = StupidPubSub()
         super().__init__(namespace, parent_logger)
 
     def _broker_reachable(self) -> bool:
         return True
 
-    async def _publish(
+    def _publish(
             self,
             namespace: str,
             payload: dict,
             ttl: int | None = None
     ) -> int:
-        return await self.pubsub.publish(namespace, payload)
+        return self.pubsub.publish(namespace, payload)
 
-    async def _listen(self):
+    def _listen(self):
         self.pubsub.subscribe(self.namespace)
-        while True:
-            try:
-                async for message in self.pubsub.listen():
-                    yield message
-            except Exception as e:
-                self.logger.exception(
-                    f"Error while reading from queue. Error msg: {e.args}"
-                )
-
-    async def initialize(self) -> None:
-        pass
+        for message in self.pubsub.listen():
+            yield message
